Cleanup code, and fix some things
